requirements.txt
gradio>=4.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
pymupdf>=1.23.0
python-docx>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
